pyproject.toml
[project]
name = "lab01-part2-nusp"
version = "0.1.0"
description = "Pipeline de ingestão de dados containerizado"
authors = [
    { name = "Gustavo" }
]
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "deltalake>=1.5.0",
    "great-expectations==0.18",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "polars>=1.39.3",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pydantic>=1.10.9,<2.0",
    "pyspark==3.5.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
